refactor: split setupDashClient into browser-safe core and Node wrapper
Extract createClient, IdentityKeyManager, AddressKeyManager, dip13KeyPath, and
KEY_SPECS into setupDashClient-core.mjs so browser apps can import them without
pulling in dotenv or process.env. setupDashClient.mjs keeps the Node-only
convenience layer and re-exports the core API so existing tutorial imports
continue to work unchanged.
